Outros tratamentos e Formatadores de resposta

+++ src/ApiFuncional/Controllers/ProdutosController.cs
@@ -17,23 +17,52 @@ namespace ApiFuncional.Controllers
         }
 
         [HttpGet]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<ActionResult<IEnumerable<Produto>>> GetProdutos() //Task de um actionResult que retorna uma coleção de Produto
         {
+            if (_context.Produtos == null)//tratamento
+            {
+                return NotFound();
+            }
+
             return await _context.Produtos.ToListAsync(); //Retorna um Ok Result por padrão - 200
         }
 
         [HttpGet("{id:int}")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<ActionResult<Produto>> GetProduto(int id)
-        { 
+        {
+            if (_context.Produtos == null)//tratamento
+            {
+                return NotFound();
+            }
+
             var produto = await _context.Produtos.FindAsync(id);
+
+            if (produto == null)//tratamento
+            {
+                return NotFound();
+            }
 
             return produto;
         }
 
         [HttpPost]
+        [ProducesResponseType(StatusCodes.Status201Created)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesDefaultResponseType]
         public async Task<ActionResult<Produto>> PostProduto(Produto produto)
         {
-            if (!ModelState.IsValid)//
+            if (_context.Produtos == null)//tratamento
+            {
+                return Problem("Erro ao criar um produto, contate o suporte!");//poderia colcoar um log de erro
+            }
+
+            if (!ModelState.IsValid)//validação
             {
                 //return BadRequest(ModelState); - Validações sem mais descrições de erro
 
@@ -52,27 +81,66 @@ namespace ApiFuncional.Controllers
         }
 
         [HttpPut("{id:int}")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<IActionResult> PutProduto(int id, Produto produto) 
         {
             if (id != produto.Id) return BadRequest();//Se o id informado na rota é o mesmo do objeto informado, caso não retorna ...
 
             if (!ModelState.IsValid) return ValidationProblem(ModelState);//Se model state for invalida, retorna..
 
-            _context.Produtos.Update(produto);
-            await _context.SaveChangesAsync();
+            _context.Entry(produto).State = EntityState.Modified;//update com tratamento de erro p/ erro do ef core, atachamento de memória
+
+            try
+            {
+                await _context.SaveChangesAsync();
+            }
+            catch (DbUpdateConcurrencyException)//tratamento de concorriência no banco 
+            {
+                if (!ProdutoExists(id))
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    throw;
+                }
+            }
+
+            //_context.Produtos.Update(produto);
+            //await _context.SaveChangesAsync();
 
             return NoContent();//204
         }
 
         [HttpDelete("{id:int}")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<IActionResult> DeleteProduto(int id) 
         {
+            if (_context.Produtos == null)//tratamento
+            {
+                return NotFound();
+            }
+            
             var produto = await _context.Produtos.FindAsync(id);
+
+            if (produto == null)//tratamento
+            {
+                return NotFound();
+            }
 
             _context.Produtos.Remove(produto);
             await _context.SaveChangesAsync();
 
             return NoContent();//204
+        }
+        private bool ProdutoExists(int id)
+        {
+            return (_context.Produtos?.Any(e => e.Id == id)).GetValueOrDefault();
         }
     }
 }
